feat: add new copy action
