+++ apps/builder/src/utils/copyManager.ts
@@ -5,6 +5,9 @@ import {
   ActionItem,
 } from "@/redux/currentApp/action/actionState"
 import store from "@/store"
+import { onCopyActionItem } from "@/page/App/components/Actions/api"
+import { componentsActions } from "@/redux/currentApp/editor/components/componentsSlice"
+import { DisplayNameGenerator } from "@/utils/generators/generateDisplayName"
 
 export class CopyManager {
   static currentCopyComponentNodes: ComponentNode[] | null = null
@@ -24,8 +27,6 @@ export class CopyManager {
         break
       case "none":
         break
-      case "inspect":
-        break
       case "components":
         break
       default:
@@ -35,20 +36,68 @@ export class CopyManager {
 
   static paste() {
     switch (FocusManager.getFocus()) {
-      case "none":
-        break
       case "dataWorkspace_action":
+      case "action":
+        if (this.currentCopyAction != null) {
+          onCopyActionItem(this.currentCopyAction)
+        }
         break
       case "dataWorkspace_component":
-        break
-      case "action":
-        break
       case "canvas":
+        if (this.currentCopyComponentNodes != null) {
+          store.dispatch(
+            componentsActions.addComponentReducer(
+              this.currentCopyComponentNodes.map((node) => {
+                return this.copyComponent(node)
+              }),
+            ),
+          )
+        }
         break
-      case "inspect":
+      case "none":
         break
       case "components":
         break
+    }
+  }
+
+  static copyComponent(
+    node: ComponentNode,
+    offsetX?: number,
+    offsetY?: number,
+  ): ComponentNode {
+    const newNode = {
+      ...node,
+      displayName: DisplayNameGenerator.generateDisplayName(
+        node.type,
+        node.showName,
+      ),
+      x: (offsetX ?? 0) + node.x,
+      y: (offsetY ?? 0) + node.y,
+    } as ComponentNode
+    this.changeComponentNodeChildrenDisplayName(newNode)
+    return newNode
+  }
+
+  private static changeComponentNodeChildrenDisplayName(
+    componentNode: ComponentNode,
+  ) {
+    if (
+      componentNode.childrenNode != null &&
+      componentNode.childrenNode.length > 0
+    ) {
+      componentNode.childrenNode.forEach((node) => {
+        node = {
+          ...node,
+          displayName: DisplayNameGenerator.generateDisplayName(
+            node.type,
+            node.showName,
+          ),
+        } as ComponentNode
+        if (node.childrenNode != null && node.childrenNode.length > 0) {
+          this.changeComponentNodeChildrenDisplayName(node)
+        }
+      })
     }
   }
 }
